Spring Boot 3.5.16 was the final OSS release of its line; the CVSS 9.8/9.1 CVEs
in spring-core/spring-web and spring-security are fixed only in Framework 7 /
Security 7. Java stays on 17.

+++ webhook-platform-worker/src/main/java/com/webhook/platform/worker/service/OrderingBufferService.java
@@ -163,7 +163,7 @@ public class OrderingBufferService {
         Long advanced = script.eval(
                 RScript.Mode.READ_WRITE,
                 cursorCasScript,
-                RScript.ReturnType.INTEGER,
+                RScript.ReturnType.LONG,
                 Collections.singletonList(key),
                 authoritative,
                 deliveredSeqTtl.toMillis());
